+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unpublished](https://github.com/metreeca/core/compare/v0.9.21...HEAD)
+## [Unpublished](https://github.com/metreeca/core/compare/v0.9.22...HEAD)
+
+## [0.9.22](https://github.com/metreeca/core/compare/v0.9.21...v0.9.22) - 2026-08-28
 
 ### Added